requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
authmanager @ file:///C:/projects/authmodule/AuthManager
billiard==3.6.4.0
blinker==1.7.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.7.1
flake8==4.0.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flower==1.0.0
greenlet==1.1.2
h11==0.13.0
humanize==4.1.0
identify==2.5.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.6.1
more-itertools==9.0.0
multidict==6.0.2
netaddr==0.10.1
nickname==0.0.0.7
nodeenv==1.6.0
numpy==1.23.1
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
pandas==1.4.3
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.12.0
PyJWT==2.6.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pyTelegramBotAPI==4.5.1
pytest==7.2.2
pytest-dependency==0.5.1
python-dateutil==2.8.2
python-environ==0.4.54
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
redis==5.0.1
requests==2.31.0
requests-toolbelt==0.9.1
resident==0.0.0.1
rich==12.4.1
selenium==4.1.5
sentry-sdk==1.39.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
SQLAlchemy-serializer==1.4.1
swagger-ui-py==23.8.23
toml==0.10.2
tomli==2.0.1
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.18
urllib3-secure-extra==0.1.0
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
Werkzeug==2.0.0
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
yookassa==3.0.0
zipp==3.8.0
retry==0.9.2
aiohttp~=3.9.5
pypdf==4.2.0
